+++ financial-connections/src/main/java/com/stripe/android/financialconnections/repository/FinancialConnectionsConsumerSessionRepository.kt
@@ -3,9 +3,11 @@ package com.stripe.android.financialconnections.repository
 import com.stripe.android.core.Logger
 import com.stripe.android.core.frauddetection.FraudDetectionDataRepository
 import com.stripe.android.financialconnections.FinancialConnectionsSheet.ElementsSessionContext
+import com.stripe.android.financialconnections.FinancialConnectionsSheet.ElementsSessionContext.BillingDetails
 import com.stripe.android.financialconnections.domain.IsLinkWithStripe
 import com.stripe.android.financialconnections.repository.api.FinancialConnectionsConsumersApiService
 import com.stripe.android.financialconnections.repository.api.ProvideApiRequestOptions
+import com.stripe.android.financialconnections.utils.toConsumerBillingAddressParams
 import com.stripe.android.model.AttachConsumerToLinkAccountSession
 import com.stripe.android.model.ConsumerPaymentDetails
 import com.stripe.android.model.ConsumerPaymentDetailsCreateParams
@@ -57,12 +59,14 @@ internal interface FinancialConnectionsConsumerSessionRepository {
     suspend fun createPaymentDetails(
         bankAccountId: String,
         consumerSessionClientSecret: String,
+        billingDetails: BillingDetails?,
     ): ConsumerPaymentDetails
 
     suspend fun sharePaymentDetails(
         paymentDetailsId: String,
         consumerSessionClientSecret: String,
         expectedPaymentMethodType: String,
+        billingPhone: String?,
     ): SharePaymentDetails
 
     companion object {
@@ -200,12 +204,15 @@ private class FinancialConnectionsConsumerSessionRepositoryImpl(
 
     override suspend fun createPaymentDetails(
         bankAccountId: String,
-        consumerSessionClientSecret: String
+        consumerSessionClientSecret: String,
+        billingDetails: BillingDetails?,
     ): ConsumerPaymentDetails {
         return consumersApiService.createPaymentDetails(
             consumerSessionClientSecret = consumerSessionClientSecret,
             paymentDetailsCreateParams = ConsumerPaymentDetailsCreateParams.BankAccount(
                 bankAccountId = bankAccountId,
+                billingAddress = billingDetails?.toConsumerBillingAddressParams(),
+                billingEmailAddress = billingDetails?.email,
             ),
             requestSurface = requestSurface,
             requestOptions = provideApiRequestOptions(useConsumerPublishableKey = true),
@@ -215,7 +222,8 @@ private class FinancialConnectionsConsumerSessionRepositoryImpl(
     override suspend fun sharePaymentDetails(
         paymentDetailsId: String,
         consumerSessionClientSecret: String,
-        expectedPaymentMethodType: String
+        expectedPaymentMethodType: String,
+        billingPhone: String?,
     ): SharePaymentDetails {
         val fraudDetectionData = fraudDetectionDataRepository.getCached()?.params.orEmpty()
 
@@ -223,6 +231,7 @@ private class FinancialConnectionsConsumerSessionRepositoryImpl(
             consumerSessionClientSecret = consumerSessionClientSecret,
             paymentDetailsId = paymentDetailsId,
             expectedPaymentMethodType = expectedPaymentMethodType,
+            billingPhone = elementsSessionContext?.billingDetails?.phone?.takeIf { it.isNotBlank() },
             requestSurface = requestSurface,
             requestOptions = provideApiRequestOptions(useConsumerPublishableKey = false),
             extraParams = fraudDetectionData,