+++ link/src/test/java/com/stripe/android/link/analytics/FakeLinkEventsReporter.kt
@@ -30,6 +30,10 @@ internal open class FakeLinkEventsReporter : LinkEventsReporter {
         throw NotImplementedError()
     }
 
+    override fun on2FAStartFailure() {
+        throw NotImplementedError()
+    }
+
     override fun onPopupShow() {
         throw NotImplementedError()
     }