+++ payments-core/src/main/java/com/stripe/android/BasicIntegrationDeprecationWarning.kt
@@ -1,5 +1,0 @@
-package com.stripe.android
-
-internal const val BASIC_INTEGRATION_DEPRECATION_WARNING = "Please use Mobile Payment Element instead. If you " +
-    "are already using Basic Integration, learn how to migrate here:" +
-    " https://docs.stripe.com/payments/mobile/migrating-to-mobile-payment-element-from-basic-integration"