+++ payments-core/src/main/java/com/stripe/android/payments/bankaccount/domain/CreateFinancialConnectionsSession.kt
@@ -4,6 +4,7 @@ import com.stripe.android.core.networking.ApiRequest
 import com.stripe.android.model.CreateFinancialConnectionsSessionForDeferredPaymentParams
 import com.stripe.android.model.CreateFinancialConnectionsSessionParams
 import com.stripe.android.model.FinancialConnectionsSession
+import com.stripe.android.model.LinkMode
 import com.stripe.android.model.PaymentIntent
 import com.stripe.android.model.SetupIntent
 import com.stripe.android.model.VerificationMethodParam
@@ -89,6 +90,7 @@ internal class CreateFinancialConnectionsSession @Inject constructor(
         customerId: String?,
         onBehalfOf: String?,
         hostedSurface: String?,
+        linkMode: LinkMode?,
         amount: Int?,
         currency: String?
     ): Result<FinancialConnectionsSession> {
@@ -102,6 +104,7 @@ internal class CreateFinancialConnectionsSession @Inject constructor(
                 verificationMethod = VerificationMethodParam.Automatic,
                 customer = customerId,
                 onBehalfOf = onBehalfOf,
+                linkMode = linkMode,
                 amount = amount,
                 currency = currency,
             ),
@@ -122,7 +125,18 @@ internal class CreateFinancialConnectionsSession @Inject constructor(
                     clientSecret = clientSecret,
                     customerName = name,
                     customerEmailAddress = email,
-                    hostedSurface = hostedSurface
+                    hostedSurface = hostedSurface,
+                    linkMode = null,
+                )
+            }
+
+            is CollectBankAccountConfiguration.USBankAccountInternal -> {
+                CreateFinancialConnectionsSessionParams.USBankAccount(
+                    clientSecret = clientSecret,
+                    customerName = name,
+                    customerEmailAddress = email,
+                    hostedSurface = hostedSurface,
+                    linkMode = elementsSessionContext?.linkMode,
                 )
             }
 
@@ -130,7 +144,8 @@ internal class CreateFinancialConnectionsSession @Inject constructor(
                 CreateFinancialConnectionsSessionParams.InstantDebits(
                     clientSecret = clientSecret,
                     customerEmailAddress = email,
-                    hostedSurface = hostedSurface
+                    hostedSurface = hostedSurface,
+                    linkMode = elementsSessionContext?.linkMode,
                 )
             }
         }