+++ paymentsheet/src/test/java/com/stripe/android/customersheet/utils/CustomerSheetHacksTest.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.testing.TestLifecycleOwner
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
+import com.stripe.android.common.coroutines.Single
 import com.stripe.android.customersheet.CustomerSheetIntegration
 import com.stripe.android.customersheet.FakeCustomerAdapter
 import com.stripe.android.customersheet.data.CustomerAdapterDataSource
@@ -15,7 +16,6 @@ import com.stripe.android.customersheet.data.CustomerSessionSavedSelectionDataSo
 import com.stripe.android.customersheet.util.CustomerSheetHacks
 import com.stripe.android.isInstanceOf
 import com.stripe.android.paymentsheet.ExperimentalCustomerSessionApi
-import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.test.runTest
 import kotlinx.coroutines.withTimeoutOrNull
 import org.junit.After
@@ -84,7 +84,7 @@ class CustomerSheetHacksTest {
         assertThat(CustomerSheetHacks.savedSelectionDataSource.awaitWithTimeout()).isNull()
     }
 
-    private suspend fun <T> Deferred<T>.awaitWithTimeout(): T? {
+    private suspend fun <T> Single<T>.awaitWithTimeout(): T? {
         return withTimeoutOrNull(1.seconds) { await() }
     }
 }